Reference the Microsoft Azure Well-Architected Framework
This week, we introduced the Microsoft Azure Well-Architected Framework.  There's a lot of overlap there with what we have here, but ultimately as that is the public direction for architecture guidance, I feel we need to follow that lead.  Some of the content we were referencing before is subsumed into the new framework.