Orchestrator references
Add references to other container orchestrators besides Service Fabric supported in Azure.  Many customers in particular will want to talk about AKS so it's good to avoid looking like we're too narrowly-focused in our view of container orchestration.

This is a small piece of a larger concern with this document, which is that it's very Service Fabric heavy when a lot of customers won't be interested in Service Fabric for microservices in 2020 - there's Azure Spring Cloud (https://azure.microsoft.com/en-us/services/spring-cloud/) for example, which Java customers are much more likely to be interested in.